License update
After some consideration, I prefer the Apache 2.0  OR MIT convention commonly seen in the Rust ecosystem.

Free (as in Libre) Software is important to me, but it is more important to me to write code that gets used. I'd still be credited under these licenses, so I still get to know that my code has actually made it somewhere it can be useful.